solution/qualityserver/requirements.txt
python-dotenv
Flask==2.3.2
prometheus-client==0.14.1

